cmd/get: resolve and download dependent mods
Abstract out mod filtering and sorting, with better flexibility to add
more filters and togglable options in the future.

This uses a primitive Goroutine and WaitGroup implementation to schedule
and run dependent mod resolution in parallel- pending abstraction and
rewrite. All dependent mods are listed and queued to be downloaded along
with the base mod. This does not provide any async download
functionality yet, but that is planned.